refactor: refine comments components
